integrate new immnodes with rest of ir

+++ src/storage/fixed_iterator.cpp
@@ -18,7 +18,7 @@ FixedIterator::FixedIterator(std::string name, const Expr& tensor, int level,
                      Int());
   idxVar = Var::make(idxVarName,Int());
 
-  this->fixedSize = (int)fixedSize;
+  this->fixedSize = (long long)fixedSize;
 }
 
 bool FixedIterator::isDense() const {
@@ -50,7 +50,7 @@ Expr FixedIterator::getIteratorVar() const {
 }
 
 Expr FixedIterator::begin() const {
-  return 0;
+  return (long long) 0;
 }
 
 Expr FixedIterator::end() const {
@@ -80,7 +80,7 @@ ir::Expr FixedIterator::getIdxArr() const {
 }
 
 ir::Stmt FixedIterator::initStorage(ir::Expr size) const {
-  return Block::make({Allocate::make(getPtrArr(), 1),
+  return Block::make({Allocate::make(getPtrArr(), (long long) 1),
                       Allocate::make(getIdxArr(), size)});
 }
